move feedback actions to components
